% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-core.R
\name{as.matrix.delarr}
\alias{as.matrix.delarr}
\title{Materialise a delayed matrix as a base matrix}
\usage{
\method{as.matrix}{delarr}(x, ...)
}
\arguments{
\item{x}{A \code{delarr}.}

\item{...}{Passed to \code{collect()}.}
}
\value{
A base matrix containing the realised data.
}
\description{
Materialise a delayed matrix as a base matrix
}
