% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_axis_chunk}
\alias{assign_axis_chunk}
\title{Assign a chunk into an array along one axis}
\usage{
assign_axis_chunk(x, block, axis, positions)
}
\description{
Assign a chunk into an array along one axis
}
\keyword{internal}
