% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{axis_scale}
\alias{axis_scale}
\title{Scale (and optionally center) along an axis}
\usage{
axis_scale(x, axis, center = TRUE, scale = TRUE, na.rm = FALSE)
}
\description{
Scale (and optionally center) along an axis
}
\keyword{internal}
