% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{colMeans2}
\alias{colMeans2}
\title{Column means for delayed matrices}
\usage{
colMeans2(x, ...)
}
\arguments{
\item{x}{An object for which row means should be computed.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Typically a numeric vector of column means.
}
\description{
Generic counterpart to \code{matrixStats::colMeans2()}. Methods are provided for
\code{delarr} objects, but packages can extend the generic for their own delayed
types.
}
\examples{
mat <- matrix(1:12, nrow = 3, ncol = 4)
darr <- delarr(mat)

# Compute column means lazily
colMeans2(darr)

# Compare with base R
colMeans(mat)
}
