% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-extended.R
\name{d_aperm}
\alias{d_aperm}
\title{Permute dimensions of a delayed array}
\usage{
d_aperm(x, perm = rev(seq_along(dim(x))), chunk_size = NULL)
}
\arguments{
\item{x}{A \code{delarr}.}

\item{perm}{A permutation of \code{seq_along(dim(x))}.}

\item{chunk_size}{Optional chunk size used for internal pulls.}
}
\value{
A permuted \code{delarr}.
}
\description{
Permute dimensions of a delayed array
}
