% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-extended.R
\name{d_matmul}
\alias{d_matmul}
\title{Delayed matrix multiplication}
\usage{
d_matmul(x, y, chunk_size = NULL)
}
\arguments{
\item{x}{A \code{delarr} or base matrix.}

\item{y}{A \code{delarr} or base matrix.}

\item{chunk_size}{Optional chunk size used during block pulls.}
}
\value{
A \code{delarr} representing \code{\%*\%}.
}
\description{
Delayed matrix multiplication
}
