% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-extended.R
\name{d_transpose}
\alias{d_transpose}
\title{Transpose a delayed matrix}
\usage{
d_transpose(x, chunk_size = NULL)
}
\arguments{
\item{x}{A \code{delarr}.}

\item{chunk_size}{Optional chunk size used for internal pulls.}
}
\value{
A transposed \code{delarr}.
}
\description{
Transpose a delayed matrix
}
