% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_axis_chunk}
\alias{extract_axis_chunk}
\title{Extract a sub-array along one axis}
\usage{
extract_axis_chunk(x, axis, positions)
}
\description{
Extract a sub-array along one axis
}
\keyword{internal}
