% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-core.R
\name{new_delarr}
\alias{new_delarr}
\title{Internal constructor for \code{delarr}}
\usage{
new_delarr(seed, ops = list())
}
\arguments{
\item{seed}{A \code{delarr_seed} backend.}

\item{ops}{A list of deferred operations.}
}
\description{
Internal constructor for \code{delarr}
}
\keyword{internal}
