% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-core.R
\name{print.delarr}
\alias{print.delarr}
\title{Pretty-print a delayed matrix}
\usage{
\method{print}{delarr}(x, ...)
}
\arguments{
\item{x}{A \code{delarr}.}

\item{...}{Unused.}
}
\value{
The original object, invisibly.
}
\description{
Pretty-print a delayed matrix
}
