% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-observe.R
\name{profile_collect}
\alias{profile_collect}
\title{Profile \code{collect()} runtime}
\usage{
profile_collect(x, reps = 3L, ...)
}
\arguments{
\item{x}{A \code{delarr}.}

\item{reps}{Number of repetitions.}

\item{...}{Additional arguments forwarded to \code{collect()}.}
}
\value{
An object of class \code{delarr_profile}.
}
\description{
Profile \code{collect()} runtime
}
