% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-verbs.R
\name{d_map}
\alias{d_map}
\title{Apply an elementwise transformation lazily}
\usage{
d_map(x, f)
}
\arguments{
\item{x}{A \code{delarr}.}

\item{f}{A function or formula suitable for \code{rlang::as_function()}.}
}
\value{
A \code{delarr} representing the transformation.
}
\description{
Apply an elementwise transformation lazily
}
