% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-verbs.R
\name{d_map2}
\alias{d_map2}
\title{Apply a binary elementwise transformation lazily}
\usage{
d_map2(x, y, f)
}
\arguments{
\item{x}{A \code{delarr}.}

\item{y}{Another \code{delarr}, matrix, or numeric vector/scalar.}

\item{f}{A function or formula combining two arguments.}
}
\value{
A \code{delarr} representing the fused binary operation.
}
\description{
Apply a binary elementwise transformation lazily
}
