% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-verbs.R
\name{d_zscore}
\alias{d_zscore}
\title{Z-score a delayed matrix}
\usage{
d_zscore(x, dim = c("rows", "cols"), na.rm = FALSE)
}
\arguments{
\item{x}{A \code{delarr}.}

\item{dim}{Dimension over which to compute the z-score.}

\item{na.rm}{Logical; remove missing values when computing statistics.}
}
\value{
A \code{delarr} with the z-score applied lazily.
}
\description{
Equivalent to centering and scaling with unit variance.
}
