% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-backends.R
\name{delarr_mem}
\alias{delarr_mem}
\title{Create a delayed matrix from an in-memory matrix}
\usage{
delarr_mem(x)
}
\arguments{
\item{x}{A numeric or logical matrix.}
}
\value{
A \code{delarr} referencing the original matrix.
}
\description{
Create a delayed matrix from an in-memory matrix
}
