% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-core.R
\name{dim.delarr}
\alias{dim.delarr}
\title{Dimensions of a delayed matrix}
\usage{
\method{dim}{delarr}(x)
}
\arguments{
\item{x}{A \code{delarr}.}
}
\value{
An integer vector of length two.
}
\description{
Computes the realised dimensions after taking queued slice and reduce
operations into account.
}
