% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-seed.R
\name{dim.delarr_seed}
\alias{dim.delarr_seed}
\title{Dimensions for a \code{delarr_seed}}
\usage{
\method{dim}{delarr_seed}(x)
}
\arguments{
\item{x}{A \code{delarr_seed}.}
}
\value{
A two-element integer vector.
}
\description{
Dimensions for a \code{delarr_seed}
}
