% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delarr-core.R
\name{dimnames.delarr}
\alias{dimnames.delarr}
\title{Dimension names for a delayed matrix}
\usage{
\method{dimnames}{delarr}(x)
}
\arguments{
\item{x}{A \code{delarr}.}
}
\value{
A list of row and column names or \code{NULL} placeholders.
}
\description{
Dimension names for a delayed matrix
}
